[ALPS04331194] FPSGO: fix potential overflow
fix signed integer overflow

MTK-Commit-Id: 1d6aaca0da8b97bac4214360acf67bbf2458cb0d

CR-Id: ALPS04331194
Feature: FPSGO
(cherry picked from commit 3326b1b53790b15ed0d6434cb0fc6baa47960ed5)

+++ drivers/misc/mediatek/performance/fpsgo_v3/fbt/src/fbt_cpu.c
@@ -1751,7 +1751,8 @@ static unsigned long long fbt_est_loading(int cur_ts,
 		return 0;
 
 	if (cur_ts > atomic_read(&last_ts)) {
-		unsigned long long dur = cur_ts - atomic_read(&last_ts);
+		unsigned long long dur = (unsigned long long)cur_ts -
+						atomic_read(&last_ts);
 
 		return dur * obv;
 	}
